Flask_GitHub/requirements.txt
beautifulsoup4==4.9.3
Bottleneck==1.3.2
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
Flask==2.0.1
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
lightfm==1.16
mal-api==0.4.0
MarkupSafe==2.0.1
numexpr==2.7.3
numpy==1.20.3
pandas==1.3.0
pickle-mixin==1.0.2
pip==21.1.3
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
sklearn==0.0
soupsieve==2.2.1
threadpoolctl==2.2.0
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
wheel==0.36.2
wincertstore==0.2
zipp==3.5.0
